feat: update main.py and implement dupe caption check refactor: some interrogate code update: add some more logging
