+++ main.py
@@ -1,55 +1,14 @@
-from typing import List, Union
-from PIL import Image
-from clip_interrogator import Interrogator, Config
-from src.caption_tools import caption_stripper
 from src.enum_model_type import ModelType
 import os
 import time
 import typer
 import json
-
-from src.logging import create_logger, register_file_logger
+from clip_interrogator import Interrogator, Config
+from src.interrogation import create_session_dirs, interrogate_image
+from src.logging import LOGGER, register_file_logger
 
 
 APP = typer.Typer()
-LOGGER = create_logger()
-
-
-def interrogate_image(file_index: int, filepath: str, CI: Interrogator, model_type: str, session_path: Union[str, None]):
-    original_caption = None
-
-    with Image.open(filepath) as image:
-        if model_type == 'fast':
-            original_caption = CI.interrogate_fast(
-                image, max_flavors=1)  # type: ignore
-        elif model_type == 'best':
-            original_caption = CI.interrogate(
-                image, max_flavors=1)  # type: ignore
-
-    stripped_caption = caption_stripper(original_caption)
-
-    if session_path and len(session_path) > 0:
-        output_path = os.path.join(
-            session_path, stripped_caption + "." + "png")
-        LOGGER.info("Image", file_index, "Output", output_path)
-        image.save(output_path, quality=100, subsampling=0)
-        return None
-    else:
-        return {"index": file_index, "caption": stripped_caption, "filetype": ".png"}
-
-
-def create_session_dirs(output_path, session_path):
-    if not os.path.exists(output_path):
-        LOGGER.info("Creating session output folder")
-        os.mkdir(output_path)
-
-    if not os.path.exists(session_path):
-        LOGGER.info("Creating session folder")
-        os.mkdir(session_path)
-
-
-def test_callback(input):
-    return input if input in ["fast", "best"] else "fast"
 
 
 def filter_files(raw_filenames: list[str]):
@@ -59,7 +18,7 @@ def filter_files(raw_filenames: list[str]):
 
 # TODO: Write Unit Tests
 @APP.command()
-def interrogate(files: List[str], model_type: ModelType = ModelType.fast, output_files: bool = typer.Option(False), log_file: bool = typer.Option(False)):
+def interrogate(files: list[str], model_type: ModelType = ModelType.fast, output_files: bool = typer.Option(False), log_file: bool = typer.Option(False)):
     if log_file:
         register_file_logger(LOGGER)
     LOGGER.info("Warming up stable-cheese")
