+++ src/caption_tools.py
@@ -8,7 +8,7 @@ def unique_words(words_as_list):
     return result
 
 
-def caption_stripper(caption):
+def optimize_caption(caption):
     words_as_list = findall(r"[A-Za-z0-9]+", caption)
     resulting_words = unique_words(words_as_list)
     return " ".join(resulting_words)
