+++ src/logging.py
@@ -2,7 +2,6 @@ from logging import INFO, FileHandler, Logger, StreamHandler, getLogger
 import os
 from sys import stdout
 
-
 def create_logger():
     logger = getLogger()
     logger.addHandler(StreamHandler(stdout))
@@ -14,3 +13,6 @@ def register_file_logger(logger: Logger):
     stdout_log_path = os.path.join("logs", "usage.log")
     if not os.path.exists(stdout_log_path): os.mkdir(os.path.dirname(stdout_log_path));
     logger.addHandler(FileHandler(stdout_log_path))
+    
+    
+LOGGER = create_logger()
